move verifyAddChain
